+++ documentation/trick_usecase.puml
@@ -10,11 +10,19 @@ usecase "Voir un trick" as UC_View
 usecase "Commenter un trick" as UC_Comment
 usecase "Ajouter une image à un trick" as UC_AddImage
 usecase "Se connecter" as UC_Login
+usecase "Authentification requise" as UC_AuthRequired
 
 Guest --> UC_View
 User --> UC_View
 
 Guest --> UC_Login
+
+UC_Create .> UC_AuthRequired : <<extends>>
+UC_Edit .> UC_AuthRequired : <<extends>>
+UC_Delete .> UC_AuthRequired : <<extends>>
+UC_Comment .> UC_AuthRequired : <<extends>>
+UC_AddImage .> UC_AuthRequired : <<extends>>
+
 User --> UC_Create
 User --> UC_Edit
 User --> UC_Delete
