+++ src/Controller/TrickController.php
@@ -1,0 +1,135 @@
+<?php
+
+namespace App\Controller;
+
+use App\Entity\Comment;
+use App\Entity\Images;
+use App\Entity\Trick;
+use App\Form\Trick\TrickFormType;
+use App\Form\Trick\TrickHandler;
+use App\Repository\TrickRepository;
+use Doctrine\ORM\EntityManagerInterface;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Bundle\SecurityBundle\Security;
+use Symfony\Component\HttpFoundation\JsonResponse;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Routing\Attribute\Route;
+
+class TrickController extends AbstractController
+{
+
+    public function __construct(
+        private readonly TrickRepository $trickRepository,
+        private readonly TrickHandler $trickHandler,
+        private readonly EntityManagerInterface $entityManager
+    )
+    {}
+
+    #[Route('/tricks', name: 'tricks')]
+    public function tricks(): Response
+    {
+        $tricks = $this->trickRepository->findAll();
+
+        return $this->render('tricks/tricks.html.twig', [
+            'tricks' => $tricks
+        ]);
+    }
+
+    #[Route('/addTrick', name: 'addTrick')]
+    public function addTrick(Request $request) : Response {
+        $trick = new Trick();
+        $form = $this->trickHandler->prepare($trick);
+
+        if ($this->trickHandler->handle($form, $request, $trick, $this->getParameter("upload_directory"))) {
+            return $this->redirectToRoute('home.index');
+        }
+
+        return $this->render('tricks/add_trick.html.twig', [
+            'form' => $form->createView(),
+        ]);
+    }
+
+    #[Route('/trick/detail/{id}', name: "home.viewTrick")]
+    public function viewTrick(string $id, Request $request, Security $security) : Response {
+
+        $trick = $this->trickRepository->find($id);
+
+        $comments = $trick->getComments();
+
+        return $this->render('tricks/view_trick.html.twig', [
+            'trick' => $trick,
+            'comments' => $comments
+        ]);
+    }
+
+    #[Route('/trick/edit/{id}', name: "editTrick")]
+    public function editTrick(string $id, Request $request) : Response {
+        $trick = $this->trickRepository->find($id);
+
+        if (!$trick) {
+            throw $this->createNotFoundException('No trick found for id '.$id);
+        }
+
+        $form = $this->trickHandler->prepare($trick);
+
+        if ($this->trickHandler->handle($form, $request, $trick, $this->getParameter("upload_directory"), true)) {
+            return $this->redirectToRoute('home.index');
+        }
+
+        return $this->render('tricks/edit_trick.html.twig', [
+            'form'  => $form->createView()
+        ]);
+    }
+
+    #[Route('/trick/remove/{id}', name: "deleteTrick")]
+    public function deleteTrick(string $id) : Response {
+
+        $trick = $this->trickRepository->find($id);
+
+        foreach ($trick->getComments() as $comment) {
+            $trick->removeComment($comment);
+        }
+
+        $imagePath = $this->getParameter('upload_directory') . '/' . $trick->getImages();
+
+        if (file_exists($imagePath)) {
+            unlink($imagePath);
+        }
+
+
+        $this->entityManager->remove($trick);
+        $this->entityManager->flush();
+
+        return $this->redirectToRoute('home.index');
+    }
+
+    public function addComment(Request $request): JsonResponse
+    {
+        $user = $this->getUser();
+        $content = $request->request->get('content');
+        $id = $request->request->get('id');
+        $trick = $this->trickRepository->find($id);
+
+        $comment = new Comment();
+        $comment->setContent($content);
+        $comment->setTrick($trick);
+        $comment->setUsername($user->getUsername());
+
+        $this->entityManager->persist($comment);
+        $this->entityManager->flush();
+
+        $updatedComments = [];
+        foreach ($trick->getComments() as $comment) {
+            $updatedComments[] = [
+                'username' => $comment->getUsername(),
+                'content' => $comment->getContent(),
+            ];
+        }
+
+        return new JsonResponse([
+            'message' => 'Commentaire ajouté avec succès',
+            'comments' => $updatedComments,
+        ]);
+    }
+}
