+++ templates/base.html.twig
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-<html>
+<html lang="en">
     <head>
         <meta charset="UTF-8">
         <title>SnowTricks | {% block title %}{% endblock %}</title>
@@ -16,7 +16,7 @@
     <body>
         <nav class="navbar navbar-expand-lg bg-dark" data-bs-theme="dark">
             <div class="container-fluid">
-                <a class="navbar-brand" href="#">Snowtricks</a>
+                <a class="navbar-brand" href="{{ path('home.index') }}">Snowtricks</a>
                 <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarColor02" aria-controls="navbarColor02" aria-expanded="false" aria-label="Toggle navigation">
                     <span class="navbar-toggler-icon"></span>
                 </button>
@@ -28,8 +28,8 @@
                         <li class="nav-item dropdown">
                             <a class="nav-link dropdown-toggle" data-bs-toggle="dropdown" role="button" aria-haspopup="true" aria-expanded="false">Figures</a>
                             <div class="dropdown-menu">
-                                <a class="dropdown-item" href="{{ path('home.tricks') }}">Toutes les figures</a>
-                                <a class="dropdown-item" href="{{ path('home.addTrick') }}">Ajouter une figure</a>
+                                <a class="dropdown-item" href="{{ path('tricks') }}">Toutes les figures</a>
+                                <a class="dropdown-item" href="{{ path('addTrick') }}">Ajouter une figure</a>
                             </div>
                         </li>
                         {% if app.user %}
@@ -52,7 +52,7 @@
         {% block body %}{% endblock %}
 
         <footer class="footer bg-primary d-flex justify-content-center align-items-center">
-            Copyright 2023
+            Copyright 2024
         </footer>
     </body>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-C6RzsynM9kWDrMNeT87bh95OGNyZPhcTNXj1NW7RuBCsyN/o0jlpcV8Qyq46cDfL" crossorigin="anonymous"></script>
