+++ templates/trick/index.html.twig
@@ -1,0 +1,20 @@
+{% extends 'base.html.twig' %}
+
+{% block title %}Hello TrickController!{% endblock %}
+
+{% block body %}
+<style>
+    .example-wrapper { margin: 1em auto; max-width: 800px; width: 95%; font: 18px/1.5 sans-serif; }
+    .example-wrapper code { background: #F5F5F5; padding: 2px 6px; }
+</style>
+
+<div class="example-wrapper">
+    <h1>Hello {{ controller_name }}! ✅</h1>
+
+    This friendly message is coming from:
+    <ul>
+        <li>Your controller at <code>C:/Users/zitro/Desktop/WarezDev/P6/snow-tricks/src/Controller/TrickController.php</code></li>
+        <li>Your template at <code>C:/Users/zitro/Desktop/WarezDev/P6/snow-tricks/templates/trick/index.html.twig</code></li>
+    </ul>
+</div>
+{% endblock %}
